+++ docs/.vuepress/plugins.js
@@ -233,6 +233,7 @@ module.exports.communityPlugins = [
   },
   {
     name: "hardhat-time-n-mine",
+    npmPackage: "@atixlabs/hardhat-time-n-mine",
     author: "Gonzalo Petraglia & Alan Verbner",
     authorUrl: "https://github.com/atixlabs",
     description:
@@ -301,6 +302,30 @@ module.exports.communityPlugins = [
     description:
       "Zero-config Hardhat plugin to generate documentation for all your Solidity contracts.",
     tags: ["Documentation", "Docs", "Solidity", "NatSpec"],
+  },
+  {
+    name: "hardhat-output-validator",
+    author: "Indeavr",
+    authorUrl: "https://github.com/indeavr",
+    description:
+      "Zero-config Hardhat plugin to check the output of the compiler for any problems like missing NatSpec",
+    tags: [
+      "Productivity",
+      "CI",
+      "Compiler",
+      "Docs",
+      "Solidity",
+      "NatSpec",
+      "Build",
+    ],
+  },
+  {
+    name: "@ericxstone/hardhat-blockscout-verify",
+    author: "ericxstone",
+    authorUrl: "https://github.com/ericxstone",
+    description:
+      "Hardhat plugin for solidity contract verification on Blockscout block explorer.",
+    tags: ["Blockscout", "Deployment", "Solidity", "Verification"],
   },
 ];