+++ packages/hardhat-core/src/internal/core/errors-list.ts
@@ -140,7 +140,7 @@ To learn more about how to access the Hardhat Runtime Environment from different
 The most common source of errors is trying to import the Hardhat Runtime Environment from your config or a file imported from it.
 This is not possible, as Hardhat can't be initialized while its config is being defined.
 
-You may also have accidentally imported \`hardhat\' instead of \`hardhat/config\`.
+You may also have accidentally imported \`hardhat\` instead of \`hardhat/config\`.
 
 Please make sure your config file is correct.
 
@@ -183,11 +183,11 @@ Please install Hardhat locally using npm or Yarn, and try again.`,
       number: 13,
       message: `Your Hardhat project uses typescript, but ts-node is not installed.
       
-Please run: npm \`install --save-dev ts-node\``,
+Please run: npm install --save-dev ts-node`,
       title: "ts-node not installed",
       description: `You are running a Hardhat project that uses typescript, but you haven't installed ts-node.
 
-Please run this and try again: npm install --save-dev ts-node`,
+Please run this and try again: \`npm install --save-dev ts-node\``,
       shouldBeReported: false,
     },
     TYPESCRIPT_NOT_INSTALLED: {
@@ -199,6 +199,21 @@ Please run: npm install --save-dev typescript`,
       description: `You are running a Hardhat project that uses typescript, but it's not installed.
 
 Please run this and try again: \`npm install --save-dev typescript\``,
+      shouldBeReported: false,
+    },
+    NOT_INSIDE_PROJECT_ON_WINDOWS: {
+      number: 15,
+      message: `You are not inside a project and Hardhat failed to initialize a new one.
+
+If you were trying to create a new project, please try again using Windows Subsystem for Linux (WSL) or PowerShell.
+`,
+      title:
+        "You are not inside a Hardhat project and Hardhat failed to initialize a new one",
+      description: `You are trying to run Hardhat outside of a Hardhat project, and we couldn't initialize one.
+
+If you were trying to create a new project, please try again using Windows Subsystem for Linux (WSL) or PowerShell.
+
+You can learn how to use Hardhat by reading the [Getting Started guide](../getting-started).`,
       shouldBeReported: false,
     },
   },