+++ packages/hardhat-core/test/internal/core/config/config-loading.ts
@@ -371,6 +371,25 @@ Hardhat plugin instead.`
       resetHardhatContext();
     });
 
+    it("should emit a warning if config is the empty object", function () {
+      loadConfigAndTasks(
+        {
+          config: "empty-config.js",
+        },
+        { showEmptyConfigWarning: true }
+      );
+
+      assert.equal(consoleWarnStub.callCount, 1);
+      assert.include(
+        consoleWarnStub.args[0][0],
+        "Hardhat config is returning an empty config object, check the export from the config file if this is unexpected."
+      );
+      assert.include(
+        consoleWarnStub.args[0][0],
+        "Learn more about configuring Hardhat at https://hardhat.org/config"
+      );
+    });
+
     it("should emit a warning if there's no configured solidity", function () {
       const config = loadConfigAndTasks(
         {