+++ packages/hardhat-ethers/src/internal/index.ts
@@ -4,7 +4,9 @@ import { lazyObject } from "hardhat/plugins";
 
 import {
   getContractAt,
+  getContractAtFromArtifact,
   getContractFactory,
+  getContractFactoryFromArtifact,
   getSigner,
   getSigners,
 } from "./helpers";
@@ -33,8 +35,6 @@ extendEnvironment((hre) => {
     return {
       ...ethers,
 
-      // The provider wrapper should be removed once this is released
-      // https://github.com/nomiclabs/hardhat/pull/608
       provider: providerProxy,
 
       getSigner: (address: string) => getSigner(hre, address),
@@ -42,7 +42,12 @@ extendEnvironment((hre) => {
       // We cast to any here as we hit a limitation of Function#bind and
       // overloads. See: https://github.com/microsoft/TypeScript/issues/28582
       getContractFactory: getContractFactory.bind(null, hre) as any,
+      getContractFactoryFromArtifact: getContractFactoryFromArtifact.bind(
+        null,
+        hre
+      ),
       getContractAt: getContractAt.bind(null, hre),
+      getContractAtFromArtifact: getContractAtFromArtifact.bind(null, hre),
     };
   });
 });