+++ packages/hardhat-vyper/test/fixture-projects/mixed-language/hardhat.config.js
@@ -2,7 +2,7 @@ require("../../../src/index");
 
 module.exports = {
   vyper: {
-    version: "0.1.0b10",
+    version: "0.3.0",
   },
   solidity: "0.5.15",
 };