+++ packages/hardhat-core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # hardhat
 
+## 2.8.3
+
+### Patch Changes
+
+- 9b1de8e0: Full rework of vyper plugin (#2082, #1364, #1338, #1335, #1258)
+- e2eb07cc: Use 40s as the default value for mocha's timeout and for the localhost network's timeout (#1549).
+
+## 2.8.2
+
+### Patch Changes
+
+- 2794fc00: Show a special error when project initialization fails on windows
+
+## 2.8.1
+
+### Patch Changes
+
+- 6753b930: Show warning if user doesn't export or exports empty object from Hardhat config file (issue #1490)
+- 0a5ab4f8: Fix running the `test` task multiple times in a script (issue #1720)
+
 ## 2.8.0
 
 ### Minor Changes