+++ packages/hardhat-ethers/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @nomiclabs/hardhat-ethers
 
+## 2.0.4
+
+### Patch Changes
+
+- 6afeeffe: Add equivalents in hardhat-ethers for `getContractFactory` and `getContractAt` that support passing `Artifact`, specifically `getContractFactoryFromArtifact` and `getContractAtFromArtifact` (issue #1716)
+
 ## 2.0.3
+
 ### Patch Changes
 
 - def9cbb2: Reset the hardhat-ethers provider when a snapshot is reverted (issue #1247)